Moved data sources into individual folders.
