+++ src/plugins/sources/appinsights/source.js
